Split renderer from formatting code
